refactor: adherence to normalization for clan inventory data
