+++ obearon/database/models/__init__.py
@@ -7,6 +7,15 @@ from .giveaway import GiveawayParticipant
 from .giveaway import GiveawayWinner
 from .guild_role import GuildRole
 from .verification import Verification
-from .warframe_player import WarframePlayer
+from .warframe_players import WarframePlayers
+from .warframe_player_names import WarframePlayerNames
 
-__all__ = ["GuildRole", "Verification", "WarframePlayer", "Giveaway", "GiveawayParticipant", "GiveawayWinner"]
+__all__ = [
+            "GuildRole",
+            "Verification",
+            "WarframePlayers",
+            "WarframePlayerNames",
+            "Giveaway",
+            "GiveawayParticipant",
+            "GiveawayWinner"
+           ]
