+++ obearon/database/models/warframe_players.py
@@ -11,15 +11,14 @@ from sqlalchemy.orm import mapped_column
 from obearon.database.engine import Base
 
 
-class WarframePlayer(Base):
+class WarframePlayers(Base):
     """
     Table that stores player data from guild inventory API.
     """
 
-    __tablename__ = "warframe_player"
+    __tablename__ = "warframe_players"
 
     oid: Mapped[str] = mapped_column(String, unique=True, nullable=False, primary_key=True)
-    names: Mapped[list[str]] = mapped_column(JSON, nullable=False)
     mastery_rank: Mapped[int] = mapped_column(Integer, nullable=False)
     in_clan: Mapped[bool] = mapped_column(default=True, nullable=False)
     blacklisted: Mapped[bool] = mapped_column(default=False, nullable=False)
